requirements.txt
dotenv
m3u-ipytv
pygit2
